# platform-api/src/resources/openapi.yaml
openapi: 3.1.1
info:
  title: WSO2 API Platform - Platform API
  description: |
    This document specifies a **RESTful API** for WSO2 **API Platform** - **Platform API**.
    
    ## Authentication
    
    All API endpoints (except organization registration) require authentication using JWT access tokens 
    generated by Thunder STS (Security Token Service).
    
    ### JWT Token Requirements
    
    - **Token Format**: Bearer token in Authorization header
    - **Required Claim**: `organization` - UUID of the user's organization
    - **Token Validation**: In development mode, signature validation can be skipped using `JWT_SKIP_VALIDATION=true`
    
    ### Organization Isolation
    
    All API operations are automatically scoped to the organization specified in the JWT token's 
    `organization` claim. Organization IDs are no longer required in request bodies or query parameters.

  version: 1.0.0
  contact:
    name: WSO2 API Platform Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://localhost:9243/api/v1
    description: Development server (HTTPS)
  - url: https://api.platform.com/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /llm-providers/{id}:
    get:
      summary: Get LLM provider by identifier
      description: Retrieve the complete configuration for a specific LLM provider.
      operationId: getLLMProvider
      tags:
        - LLM Providers
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
      responses:
        '200':
          description: LLM provider details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update an existing LLM provider
      description: Update the configuration of an existing LLM provider.
      operationId: updateLLMProvider
      tags:
        - LLM Providers
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMProvider'
      responses:
        '200':
          description: LLM provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProvider'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete an LLM provider
      description: Remove an LLM provider.
      operationId: deleteLLMProvider
      tags:
        - LLM Providers
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
      responses:
        '204':
          description: LLM provider deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /llm-providers/{id}/deployments:
    post:
      summary: Create and deploy a new LLM provider deployment
      description: |
        Creates an immutable deployment artifact for an LLM provider and deploys it to a specified gateway.
        Each deployment targets a single gateway. The providerId parameter is the LLM provider handle (identifier),
        not the UUID. The deployment becomes active immediately on the target gateway.
        Access is validated against the organization in the JWT token.
      operationId: deployLLMProvider
      tags:
        - LLM Provider Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
      requestBody:
        description: Deployment request with gateway ID, base reference, and metadata
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployAPIRequest'
      responses:
        '201':
          description: LLM provider deployed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '400':
          description: Bad Request. Validation errors (missing name, base, gatewayId, or invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Not Found. LLM provider, gateway, or base deployment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: Get deployments for an LLM provider
      description: |
        Retrieves all deployment artifacts for a specific LLM provider. The providerId parameter is the
        LLM provider handle (identifier), not the UUID. Supports filtering by gateway UUID and deployment status.
        Access is validated against the organization in the JWT token.
      operationId: getLLMProviderDeployments
      tags:
        - LLM Provider Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
        - $ref: '#/components/parameters/gatewayId-Q'
        - $ref: '#/components/parameters/deploymentStatus-Q'
      responses:
        '200':
          description: Deployments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /llm-providers/{id}/deployments/{deploymentId}:
    get:
      summary: Get LLM provider deployment by ID
      description: |
        Retrieves metadata for a specific LLM provider deployment artifact including status, gateway association,
        and timestamps. Access is validated against the organization in the JWT token.
      operationId: getLLMProviderDeployment
      tags:
        - LLM Provider Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
        - $ref: '#/components/parameters/deploymentId'
      responses:
        '200':
          description: Deployment metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete LLM provider deployment
      description: |
        Deletes a deployment artifact. Deletion is only allowed when the deployment is in UNDEPLOYED status.
        Access is validated against the organization in the JWT token.
      operationId: deleteLLMProviderDeployment
      tags:
        - LLM Provider Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
        - $ref: '#/components/parameters/deploymentId'
      responses:
        '204':
          description: Deployment deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /llm-providers/{id}/deployments/undeploy:
    post:
      summary: Undeploy LLM provider deployment from gateway
      description: |
        Undeploys an active LLM provider deployment, stopping it from being served on the specified gateway.
        The deployment artifact remains in the system with UNDEPLOYED status and can be restored later.
        Returns the updated deployment object with status changed to UNDEPLOYED.

        Both deploymentId and gatewayId are required for safety. The backend validates that the
        deployment's bound gateway matches the provided gatewayId to prevent unintended operations.
        Access is validated against the organization in the JWT token.
      operationId: undeployLLMProviderDeployment
      tags:
        - LLM Provider Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
        - name: deploymentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the deployment to undeploy
        - name: gatewayId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the gateway (validated against deployment's bound gateway)
      responses:
        '200':
          description: Deployment undeployed successfully. Returns the updated deployment with UNDEPLOYED status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /llm-providers/{id}/deployments/restore:
    post:
      summary: Restore a previous LLM provider deployment
      description: |
        Activates a previous deployment (ARCHIVED or UNDEPLOYED) by setting it as the current DEPLOYED version
        on the specified gateway. This operation changes the deployment state of the LLM provider on that gateway.
        The target deployment must not already be in DEPLOYED status.

        Both deploymentId and gatewayId are required for safety. The backend validates that the
        deployment's bound gateway matches the provided gatewayId to prevent unintended operations.
        Access is validated against the organization in the JWT token.
      operationId: restoreLLMProviderDeployment
      tags:
        - LLM Provider Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
        - name: deploymentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the deployment to restore (must be ARCHIVED or UNDEPLOYED)
        - name: gatewayId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the gateway (validated against deployment's bound gateway)
      responses:
        '200':
          description: Deployment restored successfully. Returns the deployment with status changed to DEPLOYED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict. Cannot restore currently deployed deployment or deployment is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token issued by Thunder STS. The token must contain an `organization` claim 
        that identifies the user's organization. All API operations are automatically scoped 
        to this organization.
        
        Example: `Authorization: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    Error:
      title: Error object returned
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: integer
          format: int64
        message:
          type: string
          description: Error message
        description:
          type: string
          description: A detailed description about the error message
...
    DeployAPIRequest:
      type: object
      required:
        - name
        - base
        - gatewayId
      properties:
        name:
          type: string
          description: Name/label for this deployment (e.g., "v1.0-prod", "hotfix-2024-01-15")
          example: "v1.0-production"
        base:
          type: string
          description: The source for the API definition. Can be "current" (latest working copy) or a deploymentId (existing deployment)
          example: "current"
        gatewayId:
          type: string
          format: uuid
          description: The target gateway UUID for this deployment
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata for the deployment (e.g., endpointUrl overrides)
    DeploymentResponse:
      type: object
      required:
        - deploymentId
        - name
        - gatewayId
        - status
        - createdAt
      properties:
        deploymentId:
          type: string
          format: uuid
          description: Unique identifier for the deployment
        name:
          type: string
          description: Name/label for this deployment
          example: "v1.0-production"
        gatewayId:
          type: string
          format: uuid
          description: UUID of the gateway
        status:
          type: string
          enum:
            - DEPLOYED
            - UNDEPLOYED
            - ARCHIVED
          description: |
            Current deployment lifecycle state:
            - DEPLOYED: Currently active on the gateway
            - UNDEPLOYED: Suspended but can be rolled back
            - ARCHIVED: Historical deployment, can be rolled back
        baseDeploymentId:
          type: string
          format: uuid
          nullable: true
          description: UUID of the base deployment this was created from
        metadata:
          type: object
          additionalProperties: true
          description: Metadata associated with the deployment
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the deployment artifact was created
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the deployment status last changed (null for ARCHIVED deployments)
...
    LLMProvider:
      type: object
      required:
        - id
        - name
        - version
        - template
        - upstream
        - accessControl
      properties:
        id:
          type: string
          description: Unique handle for the provider
          example: wso2-openai-provider
        name:
          type: string
          description: Human-readable LLM Provider name
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9\-_\. ]+$'
          example: WSO2 OpenAI Provider
        description:
          type: string
          description: Description of the LLM provider
          maxLength: 1023
          example: Primary OpenAI provider
        createdBy:
          type: string
          description: Username of the creator
          readOnly: true
          example: jane.doe
        version:
          type: string
          description: Semantic version of the LLM Provider
          pattern: '^v\d+\.\d+$'
          example: v1.0
        context:
          type: string
          description: Base path for all API routes (must start with /, no trailing slash)
          pattern: '^\/([a-zA-Z0-9_\-\/]*[^\/])?$'
          default: /
          minLength: 1
          maxLength: 20
          example: /openai
        vhost:
          type: string
          description: >
            Virtual host name used for routing. Supports standard domain names,
            subdomains, or wildcard domains. Must follow RFC-compliant hostname
            rules. Wildcards are only allowed in the left-most label (e.g., *.example.com).
          pattern: '^(\*\.|[a-zA-Z0-9](?:[a-zA-Z0-9\-]{0,61}[a-zA-Z0-9])?\.)+[a-zA-Z]{2,}$'
          minLength: 1
          maxLength: 253
          example: "api.openai"
        template:
          type: string
          description: Template name to use for this LLM Provider
          example: openai
        openapi:
          type: string
          description: OpenAPI specification (JSON or YAML) for the provider endpoint
          example: |
            openapi: 3.0.3
            info:
              title: Provider API
              version: v1.0
            paths: {}
        modelProviders:
          type: array
          description: >
            List of model providers and their models supported by this provider. For non-aggregator
            templates, only a single model provider is allowed (aggregator templates: awsbedrock,
            azureaifoundry).
          items:
            $ref: '#/components/schemas/LLMModelProvider'
        upstream:
          $ref: "#/components/schemas/Upstream"
        accessControl:
          $ref: '#/components/schemas/LLMAccessControl'
        rateLimiting:
          $ref: '#/components/schemas/LLMRateLimitingConfig'
        policies:
          type: array
          description: List of policies applied only to this operation (overrides or adds to API-level policies)
          items:
            $ref: "#/components/schemas/LLMPolicy"
        security:
          $ref: '#/components/schemas/SecurityConfig'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the resource was created
          readOnly: true
          example: "2023-10-12T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the resource was last updated
          readOnly: true
          example: "2023-10-12T10:30:00Z"
...
  responses:
    Unauthorized:
      description: Unauthorized. Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: Unauthorized
            description: Authorization header is required or token is invalid
    BadRequest:
      description: Bad Request. Invalid request or validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: Bad Request
            description: Invalid request or validation error
    NotFound:
      description: Not Found. The specified resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: Not Found
            description: The specified resource does not exist
...
    InternalServerError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            message: Internal Server Error
            description: The server encountered an internal error. Please contact
              administrator.
...